LC-sum-of-subarray-minimums [Runtime: N/A; Memory: N/A]

+++ monotonic-stack/sum-of-subarray-minimums.cpp
@@ -3,11 +3,11 @@ public:
     int sumSubarrayMins(vector<int>& arr) {
       int n = arr.size();
 
-      int sum = 0;
+      long long sum = 0;
 
       for(int i =0;i<n;i++){
         for(int j = i ;j<n;j++){
-            int mini = arr[j];
+            long long mini = arr[j];
             for(int k = i ; k<=j;k++){
                 if(arr[k]<mini) mini = arr[k];
             }
